melhoria nos arquivos de migracao, refatoracao geral
