+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules
-api/prisma/migrations
+prisma/migrations
 .idea/
